Fix coloring issue
