+++ packages/extensions/core/src/app.ts
@@ -34,10 +34,9 @@ import {
   writeBinary,
   writeString,
 } from "@azure-tools/uri";
-import { parseYAML } from "@azure-tools/yaml";
 import { omit } from "lodash";
 import { ArtifactWriter } from "./artifact-writer";
-import { Help } from "./help";
+import { printAutorestHelp } from "./commands";
 import { Artifact } from "./lib/artifact";
 import { AutoRest, IsOpenApiDocument, Shutdown } from "./lib/autorest-core";
 import { VERSION } from "./lib/constants";
@@ -211,45 +210,7 @@ async function currentMain(autorestArgs: Array<string>): Promise<number> {
   }
 
   if (context.config.help) {
-    // no fs operations on --help! Instead, format and print artifacts to console.
-    // - print boilerplate help
-    console.log("");
-    console.log("");
-    console.log(color("**Usage**: autorest `[configuration-file.md] [...options]`"));
-    console.log("");
-    console.log(color("  See: https://aka.ms/autorest/cli for additional documentation"));
-    // - sort artifacts by name (then content, just for stability)
-    const helpArtifacts = artifacts.sort((a, b) =>
-      a.uri === b.uri ? (a.content > b.content ? 1 : -1) : a.uri > b.uri ? 1 : -1,
-    );
-    // - format and print
-    for (const helpArtifact of helpArtifacts) {
-      const { result: help, errors } = parseYAML<Help>(helpArtifact.content);
-      if (errors.length > 0) {
-        for (const { message, position } of errors) {
-          console.error(color(`!Parsing error at **${helpArtifact.uri}**:__${position}: ${message}__`));
-        }
-      }
-      if (!help) {
-        continue;
-      }
-      const activatedBySuffix = help.activationScope ? ` (activated by --${help.activationScope})` : "";
-      console.log("");
-      console.log(color(`### ${help.categoryFriendlyName}${activatedBySuffix}`));
-      if (help.description) {
-        console.log(color(help.description));
-      }
-      console.log("");
-      for (const settingHelp of help.settings) {
-        const keyPart = `--${settingHelp.key}`;
-        const typePart = settingHelp.type ? `=<${settingHelp.type}>` : " "; // `[=<boolean>]`;
-        const settingPart = `${keyPart}\`${typePart}\``;
-        // if (!settingHelp.required) {
-        //   settingPart = `[${settingPart}]`;
-        // }
-        console.log(color(`  ${settingPart.padEnd(30)}  **${settingHelp.description}**`));
-      }
-    }
+    printAutorestHelp(artifacts);
   } else {
     // perform file system operations.
     await doClearFolders(protectFiles, clearFolders);
